fix up types, and display id
