+++ ui/frontend/build_src/src/api/index.ts
@@ -2,7 +2,7 @@
  * basic server health
  */
 
-import type { ImageRequest } from "../stores/imageCreateStore";
+import type { SAMPLER_OPTIONS } from "../stores/imageCreateStore";
 
 // when we are on dev we want to specifiy 9000 as the port for the backend
 // when we are on prod we want be realtive to the current url
@@ -57,6 +57,62 @@ export const toggleBetaConfig = async (branch: string) => {
  * post a new request for an image
  */
 // TODO; put hese some place better
+export interface ImageRequest {
+  session_id: string;
+  prompt: string;
+  seed: number;
+  num_outputs: number;
+  num_inference_steps: number;
+  guidance_scale: number;
+  width:
+  | 128
+  | 192
+  | 256
+  | 320
+  | 384
+  | 448
+  | 512
+  | 576
+  | 640
+  | 704
+  | 768
+  | 832
+  | 896
+  | 960
+  | 1024;
+  height:
+  | 128
+  | 192
+  | 256
+  | 320
+  | 384
+  | 448
+  | 512
+  | 576
+  | 640
+  | 704
+  | 768
+  | 832
+  | 896
+  | 960
+  | 1024;
+  // allow_nsfw: boolean
+  turbo: boolean;
+  use_cpu: boolean;
+  use_full_precision: boolean;
+  save_to_disk_path: null | string;
+  use_face_correction: null | "GFPGANv1.3";
+  use_upscale: null | "RealESRGAN_x4plus" | "RealESRGAN_x4plus_anime_6B" | "";
+  show_only_filtered_image: boolean;
+  init_image: undefined | string;
+  prompt_strength: undefined | number;
+  mask: undefined | string;
+  sampler: typeof SAMPLER_OPTIONS[number];
+  stream_progress_updates: true;
+  stream_image_progress: boolean;
+
+}
+
 export interface ImageOutput {
   data: string;
   path_abs: string | null;
@@ -65,9 +121,8 @@ export interface ImageOutput {
 
 export interface ImageReturnType {
   output: ImageOutput[];
-  request: {};
+  request: ImageRequest;
   status: string;
-  session_id: string;
 }
 
 export const MakeImageKey = "MakeImage";
