+++ ui/frontend/build_src/src/components/organisms/creationPanel/basicCreation/makeButton/index.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 import React, { useEffect, useRef } from "react";
 
-import { useImageCreate, ImageRequest } from "../../../../../stores/imageCreateStore";
+import { useImageCreate } from "../../../../../stores/imageCreateStore";
 import { useImageQueue } from "../../../../../stores/imageQueueStore";
 import {
   FetchingStates,
@@ -13,7 +13,12 @@ import { useImageDisplay } from "../../../../../stores/imageDisplayStore";
 import { v4 as uuidv4 } from "uuid";
 
 import { useRandomSeed } from "../../../../../utils";
-import { doMakeImage } from "../../../../../api";
+import {
+  ImageRequest,
+  ImageReturnType,
+  ImageOutput,
+  doMakeImage,
+} from "../../../../../api";
 import {
   MakeButtonStyle,
 } from "./makeButton.css";
@@ -21,6 +26,8 @@ import {
 import { useTranslation } from "react-i18next";
 
 import AudioDing from "../../../../molecules/audioDing";
+
+const idDelim = "_batch";
 
 export default function MakeButton() {
   const { t } = useTranslation();
@@ -49,45 +56,24 @@ export default function MakeButton() {
 
   const updateDisplay = useImageDisplay((state) => state.updateDisplay);
 
-  const hackJson = (jsonStr: string) => {
+  const hackJson = (jsonStr: string, id: string) => {
 
     // DONES't seem to be needed for the updated progress implementation
 
-    // if (jsonStr !== undefined && jsonStr.indexOf('}{') !== -1) {
-    //   // hack for a middleman buffering all the streaming updates, and unleashing them
-    //   //  on the poor browser in one shot.
-    //   //  this results in having to parse JSON like {"step": 1}{"step": 2}...{"status": "succeeded"..}
-    //   //  which is obviously invalid.
-    //   // So we need to just extract the last {} section, starting from "status" to the end of the response
-
-    //   const lastChunkIdx = jsonStr.lastIndexOf('}{')
-    //   if (lastChunkIdx !== -1) {
-    //     const remaining = jsonStr.substring(lastChunkIdx)
-    //     jsonStr = remaining.substring(1)
-    //   }
-    // }
-
     try {
 
-      // todo - used zod or something to validate this
-      interface jsonResponseType {
-        status: string;
-        request: ImageRequest;
-        output: []
-      }
-      const { status, request, output: outputs }: jsonResponseType = JSON.parse(jsonStr);
-
+      const parsed = JSON.parse(jsonStr);
+      const { status, request, output: outputs } = parsed as ImageReturnType;
       if (status === 'succeeded') {
-        outputs.forEach((output: any) => {
+        outputs.forEach((output: any, index: number) => {
 
-          const { data, seed } = output;
-
+          const { data, seed } = output as ImageOutput;
           const seedReq = {
             ...request,
             seed,
           };
-
-          updateDisplay(data, seedReq);
+          const batchId = `${id}${idDelim}-${seed}-${index}`;
+          updateDisplay(batchId, data, seedReq);
         });
       }
 
@@ -112,7 +98,7 @@ export default function MakeButton() {
       if (done) {
         removeFirstInQueue();
         setStatus(FetchingStates.COMPLETE);
-        hackJson(finalJSON);
+        hackJson(finalJSON, id);
         void dingRef.current?.play();
         break;
       }
