+++ ui/frontend/build_src/src/stores/imageCreateStore.ts
@@ -5,6 +5,9 @@ import { devtools } from "zustand/middleware";
 
 import { useRandomSeed } from "../utils";
 
+import { ImageRequest } from "../api";
+
+
 export interface ImageCreationUiOptions {
   isUseRandomSeed: boolean;
   isUseAutoSave: boolean;
@@ -21,62 +24,6 @@ export const SAMPLER_OPTIONS = [
   'dpm2_a',
   'lms',
 ] as const;
-
-export interface ImageRequest {
-  session_id: string;
-  prompt: string;
-  seed: number;
-  num_outputs: number;
-  num_inference_steps: number;
-  guidance_scale: number;
-  width:
-  | 128
-  | 192
-  | 256
-  | 320
-  | 384
-  | 448
-  | 512
-  | 576
-  | 640
-  | 704
-  | 768
-  | 832
-  | 896
-  | 960
-  | 1024;
-  height:
-  | 128
-  | 192
-  | 256
-  | 320
-  | 384
-  | 448
-  | 512
-  | 576
-  | 640
-  | 704
-  | 768
-  | 832
-  | 896
-  | 960
-  | 1024;
-  // allow_nsfw: boolean
-  turbo: boolean;
-  use_cpu: boolean;
-  use_full_precision: boolean;
-  save_to_disk_path: null | string;
-  use_face_correction: null | "GFPGANv1.3";
-  use_upscale: null | "RealESRGAN_x4plus" | "RealESRGAN_x4plus_anime_6B" | "";
-  show_only_filtered_image: boolean;
-  init_image: undefined | string;
-  prompt_strength: undefined | number;
-  mask: undefined | string;
-  sampler: typeof SAMPLER_OPTIONS[number];
-  stream_progress_updates: true;
-  stream_image_progress: boolean;
-
-}
 
 export interface ModifierPreview {
   name: string;
